I added a lightbbox
However, I realized that the lightbox is meant to hold all my other images that go with each project but I do not understand how that works when javascript undrtand that the four images are the placeholder for each image. how do I replace that when it's the same for each? 

I don't understand it doesn't make sense to me.